requirements.txt
aiofiles==24.1.0
aiogram==2.20
aiogram-utils==0.0.15
aiohappyeyeballs==2.4.3
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
bs4==0.0.2
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.6.1
eval-type-backport==0.2.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
magic-filter==1.0.12
mistralai==1.1.0
mongoengine==0.24.1
multidict==6.1.0
mypy-extensions==1.0.0
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing-extensions==4.12.2
typing-inspect==0.9.0
url-normalize==1.4.3
urllib3==2.2.3
yarl==1.15.2
